move executeUpnpBench to benchmark service
